fix: test issues and add roadmap plan for to monorepo refactor
