+++ src/utils/path-utils.ts
@@ -5,6 +5,25 @@
 import path from "path";
 
 /**
+ * Extracts the component name from a file path.
+ * Used for naming compiled output files from TypeScript sources.
+ *
+ * @param filePath - Full path to the component file (always uses Unix-style separators from glob)
+ * @returns Component name without the .ts extension
+ *
+ * @example
+ * _extractComponentName("/Users/project/src/hero.sb.ts") // => "hero.sb"
+ * _extractComponentName("C:/Users/project/src/card.sb.ts") // => "card.sb"
+ */
+export const extractComponentName = (filePath: string): string => {
+    const separator = "/"; // glob always returns unix-style paths
+    const sP = filePath.split(separator);
+    const lastElement = sP[sP.length - 1] as string;
+
+    return lastElement.replaceAll(".ts", "");
+};
+
+/**
  * Represents a file element with its name and path
  */
 export interface OneFileElement {
